+++ main.js
@@ -1,6 +1,6 @@
 // Required modules
-var menubar = require('menubar')
-var path = require("path")
+var menubar = require('menubar');
+var path = require("path");
 
 // Initialize manubar
 var opts = {
@@ -9,20 +9,23 @@ var opts = {
   icon: path.join(__dirname, '/build/IconTemplate.png'),
   resizable: false,
   movable: false,
+  alwaysOnTop: true,
   title: "Conda-menu",
   width: 305,
   height: 350
-}
+};
 
 // set menubar
-var mb = menubar(opts)
+var mb = menubar(opts);
 
 // On ready
 mb.on('ready', function ready () {
-  console.log('app is ready')
+  console.log('app is ready');
 })
 
 // After creation
 mb.on('after-create-window', function ready () {
   //mb.window.webContents.openDevTools() // debugging
+
+
 })