+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - added tooltips to icons  
 - fix form issues
 - added alert if conda installation could not be found
+- added ability to drag YAML conda file
+- added ability to import YAML file when creating new env
 
 ### 0.1.0 (2016-10-11)
 - First release of a working-ish version